--- cmd/gitlab-copy/migration.go
@@ -1,1 +1,0 @@
-package main
